SRCH-5524: LoadError: cannot load such file -- i18n/tasks/scanners/file_scanner (#1653)
* SRCH-5524: Bug - LoadError: cannot load such file -- i18n/tasks/scanners/file_scanner.

---------
