+++ config/i18n-tasks.yml.erb
@@ -1,4 +1,4 @@
-<% require './lib/i18n_jsx_scanner' %>
+<% require 'i18n_jsx_scanners/i18n_jsx_scanner' %>
 
 # i18n-tasks finds and manages missing and unused translations: https://github.com/glebm/i18n-tasks
